kauai/src/util.h: make offset() macro return long instead of int
Otherwise due to sizeof(int) being less than the pointer size, gcc emits the
following error:

kauai/src/util.h:64:29: error: cast from ‘ACR*’ to ‘int’ loses precision [-fpermissive]
   64 | #define offset(FOO, field) ((int)&((FOO *)0)->field)
kauai/src/rtxt.h:34:23: note: in expansion of macro ‘offset’
   34 |         ClearPb(this, offset(CHP, acrFore));

+++ kauai/src/util.h
@@ -61,7 +61,7 @@ typedef HCURSOR HCRS;
 #endif // WIN
 
 #define SIZEOF(foo) ((long)sizeof(foo))
-#define offset(FOO, field) ((int)&((FOO *)0)->field)
+#define offset(FOO, field) ((long)&((FOO *)0)->field)
 #define CvFromRgv(rgv) (SIZEOF(rgv) / SIZEOF(rgv[0]))
 #define BLOCK
